+++ tests/sdk/test_download.py
@@ -21,9 +21,9 @@ import substra
 from substra.sdk import Client
 
 from .. import datastore
-from .utils import mock_requests
-from .utils import mock_requests_responses
-from .utils import mock_response
+from ..utils import mock_requests
+from ..utils import mock_requests_responses
+from ..utils import mock_response
 
 
 @pytest.mark.parametrize(