+++ tests/sdk/test_get.py
@@ -20,9 +20,9 @@ from substra.sdk import models
 from substra.sdk import schemas
 
 from .. import datastore
-from .utils import mock_requests
-from .utils import mock_requests_responses
-from .utils import mock_response
+from ..utils import mock_requests
+from ..utils import mock_requests_responses
+from ..utils import mock_response
 
 
 @pytest.mark.parametrize(